add Timing superclass for SimpleTimer
to allow construting timing object from raw numbers
